migrate bedrock formatting into formats/ utils
