+++ crates/goose/src/providers/formats/mod.rs
@@ -1,3 +1,4 @@
 pub mod anthropic;
+pub mod bedrock;
 pub mod google;
 pub mod openai;
